refactor(gpui): move the layer types to effect_layer.rs
