+++ crates/gpui_apple/build.rs
@@ -73,6 +73,7 @@ mod macos_build {
         // Source files from gpui that define types used in shaders
         let gpui_src_paths = [
             gpui_dir.join("src/scene.rs"),
+            gpui_dir.join("src/effect_layer.rs"),
             gpui_dir.join("src/geometry.rs"),
             gpui_dir.join("src/color.rs"),
             gpui_dir.join("src/window.rs"),
